Create pi_estimator.py
Estimates py on a quantum circuit with n cubits. The algorithm iteratively searches for a solution to the equation f(x) - x = 0, where f is a quantum circuit parameterized by one parameter. This parameter represents an estimate of pi and is used instead of the actual value of pi in the inverse quantum fourier transform . With x = \pi, the equation f(x) = x is satisfied with equality. The error of the estimate decreases with the number of cubits.